Modularize App shell routes pages and shared UI modules (#14)
